chore: refactor to shared components and hooks
