+++ apps/studio/src/features/editing-experience/components/form-builder/renderers/controls/JsonFormsTagCategoryOptionsControl.tsx
@@ -1,28 +1,26 @@
-import type { DropResult } from "@hello-pangea/dnd"
 import type { ArrayLayoutProps, RankedTester } from "@jsonforms/core"
 import { Box, Flex, HStack, Text, VStack } from "@chakra-ui/react"
 import { DragDropContext, Draggable, Droppable } from "@hello-pangea/dnd"
 import {
   composePaths,
   createDefaultValue,
-  findUISchema,
   rankWith,
   schemaMatches,
 } from "@jsonforms/core"
 import { useJsonForms, withJsonFormsArrayLayoutProps } from "@jsonforms/react"
-import { Button } from "@opengovsg/design-system-react"
 import { get } from "lodash-es"
-import { useCallback, useMemo, useState } from "react"
-import { BiPlusCircle } from "react-icons/bi"
+import { useMemo, useState } from "react"
 import { JSON_FORMS_RANKING } from "~/constants/formBuilder"
 import { useIsUserIsomerAdmin } from "~/hooks/useIsUserIsomerAdmin"
 import { IsomerAdminRole } from "~prisma/generated/generatedEnums"
 
-import { ComplexEditorNestedDrawer } from "../../components/ComplexEditorNestedDrawer"
+import { AddItemButton } from "../../components/AddItemButton"
 import { DeleteConfirmModal } from "../../components/DeleteConfirmModal"
 import { DraggableTagButton } from "../../components/DraggableTagButton"
 import { DuplicateLabelError } from "../../components/DuplicateLabelError"
+import { NestedDrawerProvider } from "../../components/NestedDrawerProvider"
 import { useBuilderErrors } from "../../ErrorProvider"
+import { useArray } from "../../hooks/useArray"
 import { TagRowActionsMenu } from "./TagRowActionsMenu"
 import { indicesWithDuplicateLabels } from "./utils/indicesWithDuplicateLabels"
 
@@ -32,7 +30,6 @@ const JsonFormsTagCategoryOptionsArrayLayoutInner = (
   const {
     data,
     path,
-    visible,
     enabled,
     label,
     addItem,
@@ -42,8 +39,6 @@ const JsonFormsTagCategoryOptionsArrayLayoutInner = (
     arraySchema,
     schema,
     rootSchema,
-    renderers,
-    cells,
     uischemas,
     uischema,
     description,
@@ -61,10 +56,27 @@ const JsonFormsTagCategoryOptionsArrayLayoutInner = (
     label: string
     tagId?: string
   }>(null)
-  const [selectedIndex, setSelectedIndex] = useState<number>()
 
-  const isRemoveItemDisabled =
-    arraySchema.minItems !== undefined && data <= arraySchema.minItems
+  const arrayResult = useArray({
+    data,
+    path,
+    arraySchema,
+    schema,
+    rootSchema,
+    uischemas,
+    uischema,
+    removeItems,
+    moveUp,
+    moveDown,
+  })
+  const {
+    setSelectedIndex,
+    isAddItemDisabled,
+    isRemoveItemDisabled,
+    childUiSchema,
+    handleRemoveItem,
+    onDragEnd,
+  } = arrayResult
 
   const openDeleteModal = (index: number) => {
     const item = get(core?.data, composePaths(path, `${index}`)) as
@@ -83,85 +95,8 @@ const JsonFormsTagCategoryOptionsArrayLayoutInner = (
     setDeleteTarget(null)
   }
 
-  const handleRemoveItem = useCallback(
-    (path: string, index: number) => () => {
-      if (selectedIndex === undefined || !removeItems || isRemoveItemDisabled) {
-        return
-      }
-
-      removeItems(path, [index])()
-
-      if (selectedIndex === index) {
-        setSelectedIndex(undefined)
-      } else if (selectedIndex > index) {
-        setSelectedIndex(selectedIndex - 1)
-      }
-    },
-    [isRemoveItemDisabled, removeItems, selectedIndex],
-  )
-  const handleMoveItem = useCallback(
-    (path: string, originalIndex: number, newIndex: number) => {
-      if (originalIndex === newIndex || !moveDown || !moveUp) {
-        return
-      }
-
-      if (originalIndex < newIndex) {
-        for (let i = originalIndex; i < newIndex; i++) {
-          moveDown(path, i)()
-        }
-      } else {
-        for (let i = originalIndex; i > newIndex; i--) {
-          moveUp(path, i)()
-        }
-      }
-    },
-    [moveUp, moveDown],
-  )
-
-  const childUiSchema = useMemo(
-    () =>
-      findUISchema(
-        uischemas ?? [],
-        schema,
-        uischema.scope,
-        path,
-        undefined,
-        uischema,
-        rootSchema,
-      ),
-    [uischemas, schema, uischema, path, rootSchema],
-  )
-  const onDragEnd = useCallback(
-    (result: DropResult) => {
-      if (!result.destination) {
-        return
-      }
-      handleMoveItem(path, result.source.index, result.destination.index)
-    },
-    [path, handleMoveItem],
-  )
-
-  if (selectedIndex !== undefined) {
-    return (
-      <ComplexEditorNestedDrawer
-        renderers={renderers}
-        cells={cells}
-        visible={visible}
-        schema={schema}
-        uischema={childUiSchema}
-        path={composePaths(path, `${selectedIndex}`)}
-        label={label}
-        setSelectedIndex={setSelectedIndex}
-        isRemoveItemDisabled={isRemoveItemDisabled}
-        handleRemoveItem={handleRemoveItem(path, selectedIndex)}
-        selectedIndex={selectedIndex}
-        maxIndex={data - 1}
-      />
-    )
-  }
-
   return (
-    <>
+    <NestedDrawerProvider {...props} {...arrayResult}>
       {duplicateOptionIndices.size > 0 && <DuplicateLabelError noun="option" />}
       <VStack spacing={0} align="start">
         <VStack align="start" spacing="0.25rem" w="full">
@@ -169,19 +104,12 @@ const JsonFormsTagCategoryOptionsArrayLayoutInner = (
             <Text textStyle="subhead-1" flex={1}>
               {label}
             </Text>
-            <Button
+            <AddItemButton
               onClick={addItem(path, createDefaultValue(schema, rootSchema))}
-              variant="clear"
-              size="xs"
-              leftIcon={<BiPlusCircle fontSize="1.25rem" />}
-              isDisabled={
-                arraySchema.maxItems !== undefined &&
-                data >= arraySchema.maxItems
-              }
-              flexShrink={0}
+              isDisabled={isAddItemDisabled}
             >
               Add option
-            </Button>
+            </AddItemButton>
           </HStack>
           {description && (
             <Text textStyle="body-2" textColor="base.content.default">
@@ -314,7 +242,7 @@ const JsonFormsTagCategoryOptionsArrayLayoutInner = (
           onConfirm={handleConfirmDelete}
         />
       )}
-    </>
+    </NestedDrawerProvider>
   )
 }
